# api-spec.json
{"openapi":"3.0.0","paths":{"/input":{"post":{"operationId":"InputController_addInput","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InputConfigDto"}}}},"responses":{"201":{"description":"Input added successfully"},"400":{"description":"Bad request"}},"summary":"Add a new input configuration","tags":["input"]},"get":{"operationId":"InputController_getInputs","parameters":[],"responses":{"200":{"description":"List of input configurations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InputConfigDto"}}}}}},"summary":"Get all input configurations","tags":["input"]}},"/input/{id}":{"delete":{"operationId":"InputController_removeInput","parameters":[{"name":"id","required":true,"in":"path","description":"Input ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Input removed successfully"}},"summary":"Remove an input configuration","tags":["input"]}},"/input/{id}/data":{"get":{"operationId":"InputController_getInputData","parameters":[{"name":"id","required":true,"in":"path","description":"Input ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Input data history","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InputDataDto"}}}}}},"summary":"Get input data history","tags":["input"]}},"/input/{id}/latest":{"get":{"operationId":"InputController_getLatestData","parameters":[{"name":"id","required":true,"in":"path","description":"Input ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Latest input data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InputDataDto"}}}}},"summary":"Get latest input data","tags":["input"]}},"/transform":{"post":{"operationId":"TransformController_addTransform","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransformConfigDto"}}}},"responses":{"201":{"description":"Transform added successfully"},"400":{"description":"Bad request"}},"summary":"Add a new transform configuration","tags":["transform"]},"get":{"operationId":"TransformController_getTransforms","parameters":[],"responses":{"200":{"description":"List of transform configurations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformConfigDto"}}}}}},"summary":"Get all transform configurations","tags":["transform"]}},"/transform/{id}":{"delete":{"operationId":"TransformController_removeTransform","parameters":[{"name":"id","required":true,"in":"path","description":"Transform ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Transform removed successfully"}},"summary":"Remove a transform configuration","tags":["transform"]}},"/transform/{id}/data":{"get":{"operationId":"TransformController_getTransformData","parameters":[{"name":"id","required":true,"in":"path","description":"Transform ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Transform data history","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransformDataDto"}}}}}},"summary":"Get transform data history","tags":["transform"]}},"/transform/{id}/latest":{"get":{"operationId":"TransformController_getLatestTransformData","parameters":[{"name":"id","required":true,"in":"path","description":"Transform ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Latest transform data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransformDataDto"}}}}},"summary":"Get latest transform data","tags":["transform"]}},"/transform/{id}/apply":{"post":{"operationId":"TransformController_applyTransform","parameters":[{"name":"id","required":true,"in":"path","description":"Transform ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransformApplyDto"}}}},"responses":{"200":{"description":"Transformed data"}},"summary":"Apply a transform to data","tags":["transform"]}},"/output":{"post":{"operationId":"OutputController_addOutput","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutputConfigDto"}}}},"responses":{"201":{"description":"Output added successfully"},"400":{"description":"Bad request"}},"summary":"Add a new output configuration","tags":["output"]},"get":{"operationId":"OutputController_getOutputs","parameters":[],"responses":{"200":{"description":"List of output configurations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OutputConfigDto"}}}}}},"summary":"Get all output configurations","tags":["output"]}},"/output/{id}":{"delete":{"operationId":"OutputController_removeOutput","parameters":[{"name":"id","required":true,"in":"path","description":"Output ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Output removed successfully"}},"summary":"Remove an output configuration","tags":["output"]}},"/output/{id}/data":{"get":{"operationId":"OutputController_getOutputData","parameters":[{"name":"id","required":true,"in":"path","description":"Output ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Output data history","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OutputDataDto"}}}}}},"summary":"Get output data history","tags":["output"]}},"/output/{id}/latest":{"get":{"operationId":"OutputController_getLatestOutputData","parameters":[{"name":"id","required":true,"in":"path","description":"Output ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Latest output data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutputDataDto"}}}}},"summary":"Get latest output data","tags":["output"]}},"/output/{id}/send":{"post":{"operationId":"OutputController_sendData","parameters":[{"name":"id","required":true,"in":"path","description":"Output ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Data sent successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutputDataDto"}}}}},"summary":"Send data to an output","tags":["output"]}},"/output/{id}/test":{"post":{"operationId":"OutputController_testOutput","parameters":[{"name":"id","required":true,"in":"path","description":"Output ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutputTestDto"}}}},"responses":{"200":{"description":"Test result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OutputDataDto"}}}}},"summary":"Test an output with sample data","tags":["output"]}}},"info":{"title":"API Proxy","description":"Ein fein granulierter API-Proxy mit Transformationen, Delays und Loops","version":"1.0","contact":{}},"tags":[{"name":"input","description":"Input management - Polling external APIs"},{"name":"transform","description":"Transform management - Data transformation and processing"},{"name":"output","description":"Output management - Sending data to external APIs"}],"servers":[],"components":{"schemas":{"InputConfigDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the input","example":"weather-api"},"url":{"type":"string","description":"URL to poll","example":"https://api.weatherapi.com/v1/current.json"},"method":{"type":"string","description":"HTTP method to use","enum":["GET","POST","PUT","DELETE"],"example":"GET"},"headers":{"type":"object","description":"HTTP headers to send","example":{"Authorization":"Bearer YOUR_API_KEY"}},"body":{"type":"object","description":"Request body for POST/PUT requests","example":{"param":"value"}},"interval":{"type":"string","description":"Cron expression for polling interval","example":"*/30 * * * * *"},"enabled":{"type":"boolean","description":"Whether this input is enabled","example":true}},"required":["id","url","method","interval","enabled"]},"InputDataDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the data entry"},"timestamp":{"format":"date-time","type":"string","description":"Timestamp when data was fetched"},"data":{"type":"object","description":"The fetched data"},"source":{"type":"string","description":"Source URL of the data"}},"required":["id","timestamp","data","source"]},"TransformConfigDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the transform","example":"filter-temp"},"name":{"type":"string","description":"Human readable name for the transform","example":"Temperature Filter"},"type":{"type":"string","description":"Type of transformation","enum":["map","filter","reduce","custom"],"example":"filter"},"config":{"type":"object","description":"Configuration for the transform","example":{"condition":{"field":"temp_c","operator":"greater","value":20}}},"enabled":{"type":"boolean","description":"Whether this transform is enabled","example":true}},"required":["id","name","type","config","enabled"]},"TransformDataDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the transform data entry"},"timestamp":{"format":"date-time","type":"string","description":"Timestamp when transform was applied"},"inputData":{"type":"object","description":"Input data before transformation"},"outputData":{"type":"object","description":"Output data after transformation"},"transformId":{"type":"string","description":"ID of the transform that was applied"}},"required":["id","timestamp","inputData","outputData","transformId"]},"TransformApplyDto":{"type":"object","properties":{"data":{"type":"object","description":"Data to transform","example":{"temp_c":25,"condition":"sunny"}}},"required":["data"]},"OutputConfigDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the output","example":"slack-notification"},"name":{"type":"string","description":"Human readable name for the output","example":"Slack Notification"},"url":{"type":"string","description":"URL to send data to","example":"https://hooks.slack.com/services/YOUR_WEBHOOK"},"method":{"type":"string","description":"HTTP method to use","enum":["GET","POST","PUT","DELETE","PATCH"],"example":"POST"},"headers":{"type":"object","description":"HTTP headers to send","example":{"Content-Type":"application/json"}},"template":{"type":"string","description":"Template for request body with variable substitution","example":"{\"text\": \"Temperature: ${temp_c}°C, Condition: ${condition}\"}"},"enabled":{"type":"boolean","description":"Whether this output is enabled","example":true},"retryCount":{"type":"number","description":"Number of retry attempts on failure","example":3},"retryDelay":{"type":"number","description":"Delay between retry attempts in milliseconds","example":1000}},"required":["id","name","url","method","enabled"]},"OutputDataDto":{"type":"object","properties":{"id":{"type":"string","description":"Unique identifier for the output data entry"},"timestamp":{"format":"date-time","type":"string","description":"Timestamp when data was sent"},"requestData":{"type":"object","description":"Data that was sent"},"responseData":{"type":"object","description":"Response data from the target API"},"status":{"type":"string","description":"Status of the request","enum":["success","error"]},"error":{"type":"string","description":"Error message if request failed"},"outputId":{"type":"string","description":"ID of the output that was used"}},"required":["id","timestamp","requestData","status","outputId"]},"OutputTestDto":{"type":"object","properties":{"data":{"type":"object","description":"Test data to send","example":{"temp_c":25,"condition":"sunny"}}},"required":["data"]}}}}